refactor into internal and pkg
