+++ internal/handler.go
@@ -1,0 +1,79 @@
+func authLoginHandler(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "application/json")
+	username := r.FormValue("username")
+	password := r.FormValue("password")
+	// For simplicity, we'll use a hardcoded username and password
+	// This should be replaced with a proper authentication mechanism
+	if username == "user" && password == "pass" {
+		token, err := createToken(username)
+		if err != nil {
+			w.WriteHeader(http.StatusInternalServerError)
+			w.Write([]byte(`{"error": "Error generating the token"}`))
+			return
+		}
+		w.Write([]byte(fmt.Sprintf(`{"token": "%s"}`, token)))
+	} else {
+		w.WriteHeader(http.StatusUnauthorized)
+		w.Write([]byte(`{"error": "Invalid credentials"}`))
+	}
+}
+func authLogoutHandler(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "application/json")
+	// In this simple example, we'll just return a success message
+	w.Write([]byte(`{"message": "Logout successful"}`))
+}
+func productListHandler(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "application/json")
+	response, err := json.Marshal(products)
+	if err != nil {
+		w.WriteHeader(http.StatusInternalServerError)
+		w.Write([]byte(`{"error": "Internal Server Error"}`))
+		return
+	}
+	w.Write(response)
+}
+func productDetailHandler(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "application/json")
+	vars := mux.Vars(r)
+	productID, ok := vars["id"]
+	if !ok {
+		w.WriteHeader(http.StatusBadRequest)tea
+		w.Write([]byte(`{"error": "Product ID is missing"}`))
+		return
+	}
+	id, err := strconv.Atoi(productID)
+	if err != nil {
+		w.WriteHeader(http.StatusBadRequest)
+		w.Write([]byte(`{"error": "Product ID has wrong format"}`))
+		return
+	}
+	product := findProductByID(products, id)
+	if product == nil {
+		w.WriteHeader(http.StatusNotFound)
+		w.Write([]byte(`{"error": "Product not found"}`))
+		return
+	}
+	response, err := json.Marshal(product)
+	if err != nil {
+		w.WriteHeader(http.StatusInternalServerError)
+		w.Write([]byte(`{"error": "Internal Server Error"}`))
+		return
+	}
+	w.Write(response)
+}
+func checkoutPlaceOrderHandler(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Content-Type", "application/json")
+	token := r.Header.Get("Authorization")
+	if token == "" {
+		w.WriteHeader(http.StatusUnauthorized)
+		w.Write([]byte(`{"error": "Missing authorization header"}`))
+		return
+	}
+	if verifyToken(token) {
+		// In this simple example, we'll just return a success message
+		w.Write([]byte(`{"message": "Order placed successfully"}`))
+	} else {
+		w.WriteHeader(http.StatusUnauthorized)
+		w.Write([]byte(`{"error": "Invalid token"}`))
+	}
+}
